(Instruction) Modify LambdaInstruction class to integrate multiple template params. Fix tests and minor fix in Data::PrimitiveTypeArray to accept providing const arrays.
