+++ gegelatilib/include/instructions/lambdaInstruction.h
@@ -13,29 +13,20 @@ namespace Instructions {
 	* \brief Template instruction for simplifying the creation of an
 	* Instruction from a c++ lambda function.
 	*
-	* Template parameter T can be any type.
+	* Template parameters First and Rest can be any primitive type, class or
+	* const c-style 1D array.
 	*
-	* Currently, LambdaInstructions all require two operands, with the same
-	* type determined by T.
-	*
-	* The second template parameter is using the SFINAE to specialize the class
-	* implementation differently, depending on the template parameter.
+	* Each template parameter corresponds to an argument of the function given
+	* to the LambdaInstruction constructor, specifying its type.
 	*/
-	template <class T, typename Enable = void> class LambdaInstruction : public Instruction {};
-
-	/**
-	* \brief Specialization of the LambdaInstruction template class for
-	* primitive types.
-	*/
-	template <class T>
-	class LambdaInstruction <T, typename std::enable_if<std::is_fundamental<T>::value>::type> : public Instruction
-	{
+	template< typename First, typename... Rest>
+	class LambdaInstruction : public Instructions::Instruction {
 	protected:
 
 		/**
 		* \brief Function executed for this Instruction.
 		*/
-		const std::function<double(T, T)> func;
+		const std::function<double(const First, const Rest...)> func;
 
 	public:
 		/**
@@ -47,57 +38,14 @@ namespace Instructions {
 		* \brief Constructor for the LambdaInstruction.
 		*
 		* \param[in] function the c++ std::function that will be executed for
-		* this Instruction.
+		* this Instruction. The function must have the same types in its argument
+		* list as specified by the template parameters. (checked at compile time)
 		*/
-		LambdaInstruction(std::function<double(T, T)> function) : func{ function } {
-			this->operandTypes.push_back(typeid(T));
-			this->operandTypes.push_back(typeid(T));
-		};
+		LambdaInstruction(std::function<double(First, Rest...)> function) : func{ function } {
 
-		double execute(
-			const std::vector<std::reference_wrapper<const Parameter>>& params,
-			const std::vector<Data::UntypedSharedPtr>& args) const override {
-
-			if (Instruction::execute(params, args) != 1.0) {
-				return 0.0;
-			}
-
-			const T& arg1 = *(args.at(0).getSharedPointer<const T>());
-			const T& arg2 = *(args.at(1).getSharedPointer<const T>());
-			double result = this->func(arg1, arg2);
-			return result;
-		}
-	};
-
-	/**
-	* \brief Specialization of the LambdaInstruction template class for
-	* for arrays of primitive types.
-	*/
-	template <class T>
-	class LambdaInstruction < T, typename std::enable_if <std::is_array<T>::value>::type> : public Instruction
-	{
-	    using U = std::remove_all_extents_t<T>;
-	protected:
-		/**
-		* \brief Function executed for this Instruction.
-		*/
-		const std::function<double(const U*, const U*)> func;
-
-	public:
-		/**
-		* \brief delete the default constructor.
-		*/
-		LambdaInstruction() = delete;
-
-		/**
-		* \brief Constructor for the LambdaInstruction.
-		*
-		* \param[in] function the c++ std::function that will be executed for
-		* this Instruction.
-		*/
-		LambdaInstruction(std::function<double(const U*, const U*)> function) : func{ function } {
-			this->operandTypes.push_back(typeid(T));
-			this->operandTypes.push_back(typeid(T));
+			this->operandTypes.push_back(typeid(First));
+			// Fold expression to push all other types
+			(this->operandTypes.push_back(typeid(Rest)), ...);
 		};
 
 		/// Inherited from Instruction
@@ -106,16 +54,24 @@ namespace Instructions {
 				return false;
 			}
 
-			for (int i = 0; i < arguments.size(); i++) {
+			// List of expected types
+			const std::vector<std::reference_wrapper<const std::type_info>> expectedTypes{
+				// First
+				(!std::is_array<First>::value) ?
+					typeid(First) :
+					typeid(std::remove_all_extents_t<First>[]),
+				(!std::is_array<Rest>::value) ?
+					typeid(Rest) :
+					typeid(std::remove_all_extents_t<Rest>[])... };
+
+			for (auto idx = 0; idx < arguments.size(); idx++) {
 				// Argument Type
-				const std::type_info& argType = arguments.at(i).getType();
-				// Expected type.
-				// When requesting a c-style array "T[n]", the expected type is "const T[0]"
-				const std::type_info& ownType = typeid(const std::remove_all_extents_t<T>[]);
-				if (argType != ownType) {
+				const std::type_info& argType = arguments.at(idx).getType();
+				if (argType != expectedTypes.at(idx).get()) {
 					return false;
 				}
 			}
+
 			return true;
 		};
 
@@ -123,14 +79,33 @@ namespace Instructions {
 			const std::vector<std::reference_wrapper<const Parameter>>& params,
 			const std::vector<Data::UntypedSharedPtr>& args) const override {
 
-			// Check if operands and types have the right type.
 			if (Instruction::execute(params, args) != 1.0) {
 				return 0.0;
 			}
 
-			std::shared_ptr<const U[]> arg1 = (args.at(0)).getSharedPointer<const U[]>();
-			std::shared_ptr<const U[]> arg2 = (args.at(1)).getSharedPointer<const U[]>();
-			double result = this->func(arg1.get(),  arg2.get());
+			// const evaluated lambda expression are needed because type of arg will
+			// not be the same if First is an array, and if it is not. 
+			// Fort this reason, ternary operator can not be used.
+			const auto& arg1 = [&]() {
+				if constexpr (!std::is_array<First>::value) {
+					return *(args.at(0).getSharedPointer<const First>());
+				}
+				else {
+					return (args.at(0).getSharedPointer<const std::remove_all_extents_t<First>[]>()).get();
+				};
+			}();
+
+			size_t i = args.size() - 1;
+			// Using i-- as expansion seems to happen with parameters evaluated from right to left.
+			double result = this->func(arg1,
+				[&]() {
+					if constexpr (!std::is_array<Rest>::value) {
+						return *(args.at(i--).getSharedPointer<const Rest>());
+					}
+					else {
+						return (args.at(i--).getSharedPointer<const std::remove_all_extents_t<Rest>[]>()).get();
+					};
+				}()...);
 			return result;
 		};
 	};
